Add unit and system tests for ec2_inst_meta
This introduces the following conventions:

  * a file matching *_tests_run.ml can be compiled into an executable that will
    terminate with status 0 only if the tests it contains pass
  * a file matching *_tests.ml exports  tests that can be used to
    build the *_test_run files (what a "test" is is not yet defined)
  * we separate unit from system tests by defining different runners for
    them. Unit test runners match *_unit_tests_run.ml and system tests match
    *_system_tests_run.ml. Unit tests cannot depend on external resources

The tests introduced by this commit are still ad-hoc. There is no framework
ensuring certain homogeneity. Other things that will be addressed in future
development are:

  * Better error reporting. Currently a failed test doesn't provide any kind of
    feedback
  * Documentation about test conventions
  * Framework to avoid code duplication in the runners and test files

The system tests introduced by this module need to be run in an AWS instance
with the string "some test data" as user data.

Since we can run the system tests in different regions, we cannot match again
fixed strings in system tests as we do for unit tests, hence some system tests
are relaxed.